meta: remove funding info from readme
copied from the discord announcement:

I wanted to follow up on this, for the sake of transparency. Turns out that my death was... greatly exaggerated!

Thanks to clever caching, the actual traffic to the proxy servers is pretty limited, which means that I don't need to spend as much money as previously anticipated.

As a result, budget concerns are no longer a pressing issue at this time. However, upkeep costs do still come out of pocket, so I appreciate any donations you may want to make to support the project, and you can still contribute through Ko-fi or Liberapay if you feel so inclined.

Thanks for your understanding, and for using my project. Cheers :)